Fix semantic-form initialization.
semantic-form data was marked with DataState.Ready before all nested composite inputs were properly initialized, because of this it was not possible to use composite inputs in sparql persistence mode. In LDP mode the bug was visible only when some KPs created intermediate nodes with UUID (on every save new UUID was generated for intermediate nodes).

Signed-off-by: Artem Kozlov <artem@rem.sh>